SDK/Components - Moved naming-related functions to _naming.py
